refactor: DataDragonScheduler를 GameData 도메인으로 이동
